+++ components/atoms/BaseLabel.tsx
@@ -4,6 +4,6 @@ interface Props {
 
 export const BaseLabel = ({ text }: Props) => {
   return (
-    <span className="inline-block bg-black text-white text-sm px-2 py-0.5 rounded-md">{text}</span>
+    <span className="size-fit block bg-black text-white text-sm px-2 py-0.5 rounded-md">{text}</span>
   );
 };
